Some -Wconversion cleanups

+++ src/actions/angel.cpp
@@ -35,7 +35,7 @@ Angel::update ()
 {
   sprite.update ();
 
-  counter += game_speed;
+  counter += static_cast<float>(game_speed);
   pingu->set_pos(x_pos + 20 * Math::sin(counter * 3.0f), pingu->get_y() - 50.0f * 0.025f);
 
   // Out of screen, let the pingu die
